requirements.txt
notebook
opencv-python
fastapi[standard]
scikit-learn
python-multipart
pandas
numpy
tensorflow-cpu
matplotlib
Pillow
gunicorn
